Add "Posters" collection and page
The commit adds a new collection called "Posters" to the `_config.yml` file. It also adds a new page called "Posters" to the website, which displays a collection of posters. The page has a layout of "archive" and is accessible via the permalink "/poster/". The author profile is also displayed on this page.